SPM Support
Project importing SPM through mamba need the version to come from a different place than the plist, so the CFBundleShortVersionString is not a feasible place. This commit is just to solve some issue when testing this from the Core Video team, but a proper versioning strategy should be in place.