Adding actions for generating and publishing a release
The generate-release action is a manual dispatch action that takes
the version number as an input, and depending on the major version,
will checkout either main or main_1.x, then merge either develop or
develop_1.x into it, bump all the version strings in the repo, and
raise a PR against main or main_1.x.

The publish-release action runs on merge to main or main_1.x and
reads the version from the version.txt file to determine the
release tag it should publish, then publishes that release.

NOTE: both of these actions are broken in v2 of the repo until we
bring it up to date with the Swift Package Manager support changes,
as the version.txt file does not exist in develop currently.