Updating tags in section 4.4.2 of draft 16
Media or Multivariant Playlist tags are defined in section 4.4.2:
https://datatracker.ietf.org/doc/html/draft-pantos-hls-rfc8216bis-16#section-4.4.2

This commit updates the existing tags in that section to be in line with
draft 16.

The main change here is the addition of the EXT-X-DEFINE tag and
associated validators (both tag and playlist validators).

A further improvement may be to add handling for variable substitution
into mamba, as defined in section 4.3:
https://datatracker.ietf.org/doc/html/draft-pantos-hls-rfc8216bis-16#section-4.3

The difficulty there would be in implementing the IMPORT definition, as
it could not be self-contained to one parsed playlist, and would need a
way of associating another parsed multivariant playlist, or at least the
key-value pairs that were obtained from it.